Refactor ossl/fips code to use modules
Using includes was a useful hack to get started, but now that the
code is more stable move to use modules instead of includes whenever
possible.
