--- src/drbg.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "fips")]
-include! {"ossl/drbg.rs"}
-
-#[cfg(not(feature = "fips"))]
-include! {"ossl/drbg.rs"}
