+++ src/native/mod.rs
@@ -1,0 +1,11 @@
+// Copyright 2024 Simo Sorce
+// See LICENSE.txt file for terms
+
+#[cfg(not(feature = "fips"))]
+pub mod hmac;
+#[cfg(not(feature = "fips"))]
+pub mod pbkdf2;
+#[cfg(not(feature = "fips"))]
+pub mod sp800_108;
+#[cfg(not(feature = "fips"))]
+pub mod sshkdf;
